+++ src/josh/meanings/persistence.clj
@@ -1,4 +1,15 @@
 (ns josh.meanings.persistence
+  "A namespace for functions related to reading and writing datasets to various file formats.
+  
+  This namespace provides functions for loading datasets from files, writing datasets to files,
+  and converting between different file formats. It also provides a configuration map of
+  supported file formats and their associated reader and writer functions.
+  
+  Examples:
+
+  (read-dataset-seq state :filename)
+   
+  (write-dataset-seq state :filename [dataset1 dataset2 ...])"
   (:require
    [clojure.tools.logging :as log]
    [clojure.string]
@@ -42,47 +53,101 @@
   (last (clojure.string/split filename #"\.")))
 
 (defn filename->format
-  "Parses a format lookup key from a filename."
-  [filename]
-  {:post [(contains? formats %)]}
+  "Parses a format lookup key from a filename.
+  
+  The `filename` argument should be a string containing the filename to parse. It 
+  must contain a file extension that is in formats.
+  
+  Returns the format lookup key for the given filename."
+  [^String filename]
+  {:pre [(string? filename)]
+   :post [(contains? formats %)]}
   (-> filename extension keyword))
 
 (defn file?
   "Returns true if a file exists and false otherwise."
   [filename]
-  (.exists (clojure.java.io/file filename)))  
-  
+  (.exists (clojure.java.io/file filename)))
+
+
 (defn read-dataset-seq
-  "Loads the dataset at the file path in key."
-  [k-means-state key]
-  (let [filename (key k-means-state)
-        format (filename->format filename)
-        reader-fn (-> formats format :reader)]
-    (log/info "Loading" filename "with" format)
-    (reader-fn filename)))
+  "Loads the dataset at the file path in key.
+   
+  If the `filename` argument is provided, it should be a string containing the
+  file path of the dataset to load. If the `s` and `key` arguments are provided,
+  the `s` argument should be a map containing the `key` value that identifies
+  the file path of the dataset to load.
+  
+  Returns a sequence of datasets from the specified file.
+   
+  Examples:
+    
+  (read-dataset-seq '/path/to/file.csv')
+  (read-dataset-seq {:filename '/path/to/file.csv'} :filename)"
+  ([]
+   (throw (IllegalArgumentException. "Missing required argument: filename")))
+  ([^String filename] 
+   {:post [(seq? %) (every? ds/dataset? %)]}
+   (let [format (filename->format filename)
+         reader-fn (-> formats format :reader)]
+     (log/info "Loading" filename "with" format)
+     (reader-fn filename)))
+  ([^clojure.lang.IPersistentMap s ^clojure.lang.Keyword key]
+   {:pre [(map? s) (keyword? key)]}
+   (read-dataset-seq (key s))))
 
 (defn dataset-seq->column-names
+  "Returns a sequence of column names for a given dataset sequence.
+  
+  The `ds-seq` argument should be a sequence of datasets for which to
+  return the column names. The returned sequence will contain the column
+  names for the first dataset in the sequence.
+  
+  Returns a sequence of column names for the given dataset sequence."
   [ds-seq]
   (ds/column-names (first ds-seq)))
 
-(defn write-dataset
-  [k-means-state key dataset]
-  (let [filename (key k-means-state)]
-    (log/info "About to write" key "with data" dataset)
-    (log/info "Writing to" filename)
-    (ds/write! dataset filename)
-    (log/info "Finished writing " filename)))
 
+(defn write-dataset
+  "Writes a dataset to a file using the specified format.
+  
+  The `dataset` argument should be a sequence of datasets to write to the file. 
+  The file format will be determined by the file extension of the filename.
+  
+  Returns nil."
+  ([filename dataset]
+   (log/info "About to write" filename "with data" dataset)
+   (ds/write! dataset filename)
+   (log/info "Finished writing " filename))
+  ([s key dataset]
+   {:pre [(map? s) (keyword? key) (ds/dataset? dataset)]}
+   (write-dataset (key s) dataset)))
+
+(defn write-datasets
+  "Writes a sequence of datasets to filename.
+  
+  The `ds-seq` argument should be a sequence of datasets to write to the file. 
+  The file format will be determined by the file extension of the filename.
+   
+  Returns nil."
+  [filename ds-seq]
+  (let [format (filename->format filename)
+        writer-fn! (-> formats format :writer)]
+    (log/info "About to write" filename)
+    (writer-fn! filename ds-seq)
+    (log/info "Finished writing " filename)))
 
 (defn write-dataset-seq
-  [k-means-state key dataset]
-  (let [filename (key k-means-state)
-        format (filename->format filename)
-        writer-fn! (-> formats format :writer)]
-    (log/info "About to write" key "with data" dataset)
-    (log/info "Writing to" filename)
-    (writer-fn! filename dataset)
-    (log/info "Finished writing " filename)))
+  "Writes a sequence of datasets to a file using the specified format.
+  
+  The `k-means-state` argument should be a map containing the `key` value that
+  identifies the file to write to. The `ds-seq` argument should be a sequence of
+  datasets to write to the file. The file format will be determined by the file
+  extension of the filename specified in the `k-means-state` map.
+  
+  Returns nil."
+  [k-means-state key ds-seq]
+  (write-datasets (key k-means-state) ds-seq))
 
 
 (defn change-extension
@@ -90,25 +155,21 @@
   (let [desired-suffix (:suffix (formats format))]
     (clojure.string/replace filename #"(.*)\.(.*?)$" (str "$1" desired-suffix))))
 
-(defn csv-seq-filename->format-seq
-  "Converts a csv file into another file type and 
-   returns a k means object with an updated key name.
-   
-  CSV isn't a well optimized format for doing large computations.                                                        
-  Computing the min and max of each column working with optimized 
-  csv seqs can take two orders of magnitude longer than the same 
-  operations performed against arrow streams."
-  [k-means-state key]
-  (let [format (:format k-means-state)
-        input-filename (key k-means-state)
-        new-filename (change-extension input-filename format)
-        new-state (assoc k-means-state key new-filename)]
-    (log/info "Requested conversion of" input-filename "to" new-filename)
-    (when (not= input-filename new-filename)
-      (log/info "Converting" input-filename "to" new-filename)
-      (write-dataset-seq new-state key (ds-csv/csv->dataset-seq input-filename {:header-row? true}))
+
+
+
+(defn convert-file
+  "Converts a file into another file type and returns the new filename."
+  [filename format]
+  (let [new-filename (change-extension filename format)
+        reader-fn (:reader (formats (filename->format filename)))]
+    (when (not= filename new-filename)
+      (log/info "Converting" filename "to" new-filename)
+      (write-datasets new-filename (reader-fn filename))
       (log/info "Conversion completed"))
-    new-state))
+    new-filename))
+
+
 
 (defn ds-seq->rows->maps
   [ds-seq rows]