+++ src/josh/meanings/protocols/classifier.clj
@@ -1,0 +1,7 @@
+(ns josh.meanings.protocols.classifier)
+
+
+(defprotocol Classifier
+  (load-assignments [this])
+  (load-centroids [this])
+  (classify [this x]))