+++ src/josh/meanings/protocols/savable.clj
@@ -1,0 +1,4 @@
+(ns josh.meanings.protocols.savable)
+
+(defprotocol Savable
+  (save-model [this filename]))